// src/types/socket.ts
export enum SOCKET_EVENT {
  NEW_MESSAGE = 'NEW_MESSAGE',
  MESSAGE_RECEIVED = 'MESSAGE_RECEIVED',
  MESSAGE_SEEN = 'MESSAGE_SEEN',
  UPDATE_MESSAGE = 'UPDATE_MESSAGE',
  NEW_CHAT = 'NEW_CHAT',
  NEW_CONNECTION = 'NEW_CONNECTION',
  CONNECTION_ACCEPTED = 'CONNECTION_ACCEPTED',
  CONNECTION_CANCELED = 'CONNECTION_CANCELED',
}